+++ lib/cliutils/constants.rb
@@ -8,5 +8,5 @@ module CLIUtils
   SUMMARY = 'Sugary goodness for Ruby CLI apps.'
 
   # The current version of the gem
-  VERSION = '2.0.3'
+  VERSION = '2.1.0'
 end